in_opentelemetry: split the Handler class into proper files
